+++ packages/eslint-config/react.js
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: ['airbnb-typescript', './src/shared', 'prettier/react'],
+  rules: {
+    'react/react-in-jsx-scope': 'off',
+  },
+}