+++ packages/eslint-config/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+- Update to be based on airbnb's eslint config
+- Include a react config in this package (previously a separate package was available)
+
 ## 1.0.0
 
 - Initial release