+++ packages/prettier-config/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.0
+
+- Change `bracketSpacing` to `true`
+
 ## 1.0.0
 
 - Initial release